provide trait methods for converting from Txn to StateUpdates, from ProposalBlock to StateUpdates and from StateUpdate to UpdateArgs. Use Address for sender and receiver addresses throughout codebase for transactions rather than String
fix broken tests in mempool related to change in txn

fix errors in miner/test_helpers caused by changing txn

fix errors in vrrbdb and validator caused by changes in txn

fix tests related to change in txn

fix tests in block_processing and state_sync related to change in txn

modify mempool tests to ensure different txns are treated differently, in otherwords, needed two separate receiver addresses otherwise the transaction digests were exactly the same and treated as the same

ignore validator test for now until validator_core can be refactored to account for the change in the Txn struct, which now uses proper Address instead of String addresses

fix test in mempool that fails when retrieval timestamp is 1 ms different than preconfigured 'now' timestamp

fix rpc test to use correct address in both actual and mock record for accurate comparison

fix wallet_rpc_test so that the digest is repeatable and testable

+++ crates/vrrb_core/src/account.rs
@@ -27,7 +27,7 @@ pub enum AccountField {
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct UpdateArgs {
     pub address: Address,
-    pub nonce: u32,
+    pub nonce: u128,
     pub credits: Option<u128>,
     pub debits: Option<u128>,
     pub storage: Option<Option<String>>,
@@ -74,7 +74,7 @@ impl Hash for UpdateArgs {
 }
 
 
-pub type AccountNonce = u32;
+pub type AccountNonce = u128;
 
 #[derive(Clone, Default, PartialEq, Eq, Debug, Serialize, Deserialize)]
 pub struct Account {
@@ -93,7 +93,7 @@ pub struct Account {
 impl Account {
     /// Returns new, empty account.
     pub fn new(pubkey: secp256k1::PublicKey) -> Account {
-        let nonce = 0u32;
+        let nonce = 0u128;
         let credits = 0u128;
         let debits = 0u128;
         let storage = None;
